:bug: init.shを実行したときにエラーで止まるバグを修正

+++ etc/init/setup/iCloudDrive.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-[[ -e $HOME/iCloudDrive ]] || {
+if ! [[ -e $HOME/iCloudDrive ]]; then
   ln -s $HOME/Library/Mobile\ Documents/com~apple~CloudDocs $HOME/iCloudDrive &&
   echo 'Created symlink to $HOME/iCloudDrive'
-}
+fi
